.sandboxtmp/wheeltest2_7443c704e0a14886b1a882eebf6a744e/project/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polytool"
version = "0.1.0"
description = "Local-first toolchain for Polymarket analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PolyTool Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.28.0",
    "clickhouse-connect>=0.6.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
rag = [
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
]
mcp = [
    "mcp>=1.0.0",
]
simtrader = [
    "websocket-client>=1.6",
]
studio = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
all = [
    "polytool[rag,mcp,simtrader,studio,dev]",
]

[project.scripts]
polytool = "polytool.__main__:main"

[project.urls]
Homepage = "https://github.com/polymarket/polytool"
Documentation = "https://github.com/polymarket/polytool/tree/main/docs"

[tool.setuptools]
packages = [
    "polytool", "polytool.reports",
    "packages", "packages.polymarket", "packages.polymarket.rag",
    "packages.polymarket.hypotheses",
    "packages.polymarket.market_selection",
    "packages.research", "packages.research.hypotheses",
    "packages.polymarket.simtrader",
    "packages.polymarket.simtrader.studio",
    "packages.polymarket.simtrader.tape",
    "packages.polymarket.simtrader.orderbook",
    "packages.polymarket.simtrader.replay",
    "packages.polymarket.simtrader.broker",
    "tools", "tools.cli", "tools.guard",
]

[tool.setuptools.package-data]
"packages.polymarket.hypotheses" = ["hypothesis_schema_v1.json"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short -p no:cacheprovider"
norecursedirs = ["tests/pytest_tmp", "tests/.pytest_tmp", ".tmp", "pytest_tmp", ".pytest_tmp"]
markers = [
    "optional_dep: tests that require optional third-party dependencies",
]

